+++ src/main/kotlin/io/github/heyhey123/xiaojiegui/skript/elements/session/effects/EffCloseSession.kt
@@ -31,7 +31,7 @@ class EffCloseSession : Effect() {
             SkriptSyntax.effect(
                 addon,
                 EffCloseSession::class.java,
-                "close [the] [menu] [session] %menusession%"
+                "close [the] [menu] [(session|window)] %menusession%"
             )
         }
     }